+++ CHANGELOG.md
@@ -8,6 +8,9 @@
     Due to a limitation of the [`lestrrat-go/libxml2`](https://github.com/lestrrat-go/libxml2/) package,  
     `FABified` EADs still require a processing step to delete subcontainer  
     `@id` attributes.
+  - Add stringification of `<title @render=...>` attribute
+  - Fix bug in stringification of `<emph>` elements that do not have a `@render` attribute
+  
     
 #### v0.27.1
   - Bug Fix: Add closing `</a>` for `<ref>` elements